% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corenames}
\alias{corenames}
\title{R Core members}
\format{A data frame with 20 rows and two columns:
\describe{
   \item{First}{First, or given, name of the R core members}
   \item{Last}{Last, or family, name(s)}
}}
\source{
\url{https://www.r-project.org/contributors.html}, accessed 2019-11-25
}
\usage{
corenames
}
\description{
Names of R core members, scraped from the R homepage.
}
\keyword{datasets}
