% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names}
\alias{names}
\title{First names}
\format{A character vector of length four.}
\source{
Synthetic data set
}
\usage{
names
}
\description{
A vector with four common English first names.
}
\keyword{datasets}
