% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\docType{package}
\name{say}
\alias{say}
\alias{say-package}
\title{A package for saying things}
\description{
This package allows the user to say hello (for now).
}
\section{Future plans}{

Many, but currently focusing on good bye.
}

