% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{A function to say hello}
\usage{
say_hello(name = "my little friend", punct = ".")
}
\arguments{
\item{name}{Character string containing the name of the
gretee, with a reasonable default}

\item{punct}{Character to terminate the greeting with;
default is full stop (\code{.})}
}
\value{
\code{NULL}. This function is called for its side effect,
        to print a greeting to the console.
}
\description{
Specify a name and a punctuation character,
and the function will print a personalized greeting.
}
\details{
This function works by concatenating strings. Note
         that by the power of R, this function is inherently
          vectorized, see Examples.
}
\note{
Currently, this function will only greet in English.
}
\examples{
# Generic defaults
say_hello()

# Note: two very different use cases
say_hello("LETTERS")
say_hello(LETTERS, punct = ". ")

# Much flexibility
say_hello(name = c("Brian", "Fritz"), punct = "!\\n")
}
\references{
For an overview of earlier, less flexible implementations,
   see \url{https://en.wikipedia.org/wiki/\%22Hello,_World!\%22_program}
}
\seealso{
\code{\link[base]{paste}}
}
\author{
Alexander Ploner \email{Alexander.Ploner@ki.se}
}
